fixup! fixup! JAMES-2295 JMS mailqueue deadlock when a single thread issue several consume.
